+++ src/components/_variables.scss
