+++ src/components/general/contact.scss
@@ -1,5 +1,5 @@
 #contact {
-  background: #000;
+  background: rgb(0, 0, 0);
   color: rgba(255, 255, 255, 0.8);
   box-shadow: 0px -4px 20px -2px rgba(0, 0, 0, 0.5);
   padding-bottom: 4rem;
