+++ src/components/general/footer.scss
@@ -1,4 +1,4 @@
-@import './_variables';
+@import '../_variables';
 
 footer {
   padding: 3rem 0 0;
