+++ src/components/general/header.scss
@@ -1,4 +1,4 @@
-@import './_variables';
+@import '../_variables';
 
 $nav-spacing: 1.5rem;
