+++ src/components/homepage/banner.scss
@@ -1,6 +1,6 @@
 @import url('https://fonts.googleapis.com/css?family=Orbitron&display=swap');
 @import url('https://fonts.googleapis.com/css?family=Share+Tech+Mono&display=swap');
-@import "./_variables.scss";
+@import "../_variables.scss";
 #banner {
   .hr{
     background-color: $sectionSeparator;
