+++ src/components/homepage/hero.scss
@@ -1,5 +1,5 @@
 
-@import "./_variables.scss";
+@import "../_variables.scss";
 #hero {
   padding-top: 8rem;
   padding-bottom: 10rem;
@@ -16,9 +16,6 @@
         color: $primary;
         font-weight: bold;
         margin: 0;
-        @media screen and (max-width: 768px){
-          font-size: 4rem;
-        }
         .stroke{
           -webkit-text-stroke: 3px $primary;
           color: rgba($color: #000000, $alpha: 0);
@@ -32,5 +29,13 @@
   #hero{
     padding-top: 3rem;
     padding-bottom: 4rem;
+    .hero-container{
+      .container-center{
+        margin-left: 12%;
+        h1{
+          font-size: 3.5rem;
+        }
+      }
+    }
   }
 }
