+++ src/components/homepage/parallaxbox.scss
@@ -1,4 +1,4 @@
-@import './_variables';
+@import '../_variables';
 #parallaxbox{
   height: 500px;
   .parallax-banner{
