+++ src/components/homepage/rectangles.js
