+++ src/components/homepage/rectangles.scss
@@ -1,4 +1,4 @@
-@import "./_variables.scss";
+@import "../_variables.scss";
 
 
 #rectangles{
