+++ src/components/homepage/textblock.scss
@@ -1,4 +1,4 @@
-@import './_variables';
+@import '../_variables';
 
 .textblock{
   .iconsinline{
