+++ src/components/homepage/uianimation.scss
@@ -1,0 +1,88 @@
+@import '../_variables';
+
+#uianimation{
+  .bottom-right{
+    position: relative;
+    width: 10%;
+    border-radius: 10px;
+    border: 2px solid $primary;
+    overflow: hidden;
+    .right-fill{
+      position: absolute;
+      bottom: 0;
+      height: 75%;
+      width: 100%;
+      background: $primary;
+    }
+  }
+  .frame-bottom{
+    display: flex;
+    align-content: stretch;
+    height: 100%;
+    padding-top: 20px;
+
+  }
+  .frame-top{
+    height: 1rem;
+    width: 100%;
+    background: lightslategray;
+    border-radius: 5px;
+  }
+  .bottom-left{
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    width: 100%;
+    padding-right: 20px;
+    .balken{
+      width: 100%;
+      height: 1rem;
+      background: lightgray;
+      border-radius: 5px;
+    }
+  }
+
+  .frame{
+    opacity: 0.75;
+    position: relative;
+    z-index: 1;
+    width: 100%;
+    height: 200px;
+    border: 5px solid $primary;
+    border-radius: 10px;
+    padding: 20px;
+    display: flex;
+    flex-direction: column;
+    margin-top: -35%;
+  }
+  .square-container{
+    transform: translate3d(-50%, -50%, 0);
+    position: relative;
+    z-index: 2;
+    margin-left: 3px;
+    display: flex;
+    width: calc(5rem + 60px);
+    flex-wrap: wrap;
+    background: radial-gradient(42.4% 38.38% at 76.25% 6.26%,
+    rgba(226, 124, 156, 0.5) 0%, rgba(226, 124, 156, 0.1) 100%), radial-gradient(52.09% 52.09% at 31.94% 7.01%,
+    rgba(95, 195, 228, 0.5) 0%, rgba(95, 195, 228, 0.1) 100%), #000000;
+    border-radius: 20px;
+    padding: 6px;
+
+  }
+  .square{
+
+    margin: 12px;
+    height: 2.5rem;
+    width: 2.5rem;
+    background: linear-gradient(-45deg, #b9b6a7,  #eaeaea);
+    border-radius: 10px;
+  }
+
+  @media (max-width: 768px){
+    .square-container{
+        transform: translate3d(-75%,-50%,0);
+        margin: auto;
+    }
+  }
+}
