+++ src/components/pointer.scss
@@ -30,7 +30,7 @@
 }
 
 html {
-    cursor: none !important; 
+    cursor: default !important; 
 }
 
 a {
