--- src/components/sectionone.scss
@@ -1,248 +1,0 @@
-@import "./_variables.scss";
-
-#sectionone{
-  
-  padding-top: 12rem;
-  padding-bottom: 6rem;
-  
-
-  .hr{
-    background-color: $sectionSeparator;
-    height: 0.5px;
-    width: 50%;
-    margin: 10rem auto;
-  }
-  
-  .centercontent{
-    display: flex;
-    align-items: center;
-  }
-  
-  .block-margin-top{
-    position: relative;
-    z-index: 1;
-    margin-top: 6rem;
-    img{
-      border-radius: 10px;
-    }
-  }
-  
-  .forms{
-    position: absolute;
-    top: 0;
-    left: 0;
-    z-index: 0;
-    transform: translate3d(-45%, -45%, 0);
-  }
-  .bottom-right{
-    position: relative;
-    width: 10%;
-    border-radius: 10px;
-    border: 2px solid $primary;
-    overflow: hidden;
-    .right-fill{
-      position: absolute;
-      bottom: 0;
-      height: 75%;
-      width: 100%;
-      background: $primary;
-    }
-  }
-  .frame-bottom{
-    display: flex;
-    align-content: stretch;
-    height: 100%;
-    padding-top: 20px;
-
-  }
-  .frame-top{
-    height: 1rem;
-    width: 100%;
-    background: lightslategray;
-    border-radius: 5px;
-  }
-  .bottom-left{
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    width: 100%;
-    padding-right: 20px;
-    .balken{
-      width: 100%;
-      height: 1rem;
-      background: lightgray;
-      border-radius: 5px;
-    }
-  }
-  
-  .frame{
-    opacity: 0.75;
-    position: relative;
-    z-index: 1;
-    width: 100%;
-    height: 200px;
-    border: 5px solid $primary;
-    border-radius: 10px;
-    padding: 20px;
-    display: flex;
-    flex-direction: column;
-    margin-top: -35%;
-  }
-  .squareholder{
-    transform: translate3d(-50%, -50%, 0);
-    position: relative;
-    z-index: 2;
-    margin-left: 3px;
-    display: flex;
-    width: calc(5rem + 60px);
-    flex-wrap: wrap;
-    background: radial-gradient(42.4% 38.38% at 76.25% 6.26%,
-    rgba(226, 124, 156, 0.5) 0%, rgba(226, 124, 156, 0.1) 100%), radial-gradient(52.09% 52.09% at 31.94% 7.01%,
-    rgba(95, 195, 228, 0.5) 0%, rgba(95, 195, 228, 0.1) 100%), #000000;
-    border-radius: 20px;
-    padding: 6px;
-
-  }
-  .square{
-
-    margin: 12px;
-    height: 2.5rem;
-    width: 2.5rem;
-    background: linear-gradient(-45deg, #b9b6a7,  #eaeaea);
-    border-radius: 10px;
-  }
-
-  .car{
-    
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-  .left{
-    transform: rotateZ(180deg);
-    
-  }
-  .arrow{
-    width: 1.5rem;
-    
-  }
-  .car-middle{
-    width: 75%;
-    margin: 0;
-  }
-  .people{
-    margin: 0;
-  }
-  h2{
-    font-weight: 800;
-    margin-bottom: 0;
-    font-size: 2.8rem;
-  }
-  .image-middle{
-    position: relative;
-    margin: 0;
-    margin-left: 50%;
-    margin-top: -10%;
-    z-index: 2;
-
-    
-    
-  }
-  .custom-class{
-    position: relative;
-    z-index: 1;
-  }
-  .image-right{
-    position: relative;
-    float: right;
-    
-  }
-  
-  .progressring{
-    display: flex;
-    color: green;
-    justify-content: center;
-    align-items: center;
-    height: 62px;
-    p{
-      margin-bottom: 0;
-      font-size: 1.5rem;
-      font-weight: bold;
-      color: green;
-    }
-    svg{
-      position: absolute;
-      
-      circle{
-        
-      }
-    }
-  }
-  .circle-text{
-    color: white;
-    text-align: center;
-    font-weight: bold;
-    margin-top: 1rem;
-  }
-  .fullheightcolumn{
-    height: 100%;
-    display: flex;
-    flex-wrap: wrap;
-  }
-
-    .ring{
-      width: 50%;
-      margin-bottom: 2rem;
-    }
-    
-
-  .ring-column{
-    
-  }
-  .code-image{
-    
-    margin-top: -50%;
-    opacity: 0.5;
-  }
-
-  .center-col{
-    display: flex;
-    align-items: center;
-
-  }
-}
-
-@media (max-width: 768px){
-  #sectionone{
-    .image-middle{
-      margin-top: -3rem;
-      margin-left: 0;
-      float: right;
-    }
-    .block-margin-top{
-      margin-top: -3rem;
-      img{
-        width: 90%;
-      }
-    }
-    .fullheightcolumn{
-      margin-top: 4rem;
-      margin-bottom: 2rem;
-    }
-    .squareholder{
-      transform: translate3d(-75%,-50%,0);
-      margin: auto;
-    }
-    .code-image{
-      margin-left: 0;
-    }
-    .circle-text{
-      color: white;
-      text-align: center;
-      font-weight: bold;
-      font-size: 0.8rem;
-      margin-top: 1rem;
-    }
-  }
-  
-}
