+++ src/components/sites/members.scss
@@ -1,4 +1,4 @@
-@import './_variables';
+@import '../_variables';
 
 #members {
   padding: 5rem 0;
@@ -30,26 +30,25 @@
       }
     }
     .member1{
-      background: url("../images/team/lukas.jpg") center;
+      background: url("../../images/team/lukas.jpg") center;
       background-size: contain;
     }
     .member2{
-      background: url("../images/team/antonio.jpg") center;
+      background: url("../../images/team/antonio.jpg") center;
       background-size: contain;
     }
     .member3{
-      background: url("../images/team/david.jpg") center;
+      background: url("../../images/team/david.jpg") center;
       background-size: contain;
     }
     .member4{
-      background: url("../images/team/moritz.jpg") center;
+      background: url("../../images/team/moritz.jpg") center;
       background-size: contain;
     }
     .member5{
-      background: url("../images/team/oskar.jpg") center;
+      background: url("../../images/team/oskar.jpg") center;
       background-size: contain;
     }
-
     h3 {
       font-weight: 800;
       font-size: 1.4rem;
