+++ src/components/sites/project.scss
@@ -1,4 +1,4 @@
-@import "./_variables.scss";
+@import "../_variables.scss";
 
 .projects {
   padding: 4rem 0;
