+++ src/components/sites/services.scss
@@ -1,4 +1,4 @@
-@import './_variables';
+@import '../_variables';
 
 #services {
   text-align: center;
