+++ src/components/styles.scss
@@ -4,13 +4,13 @@
 * {
   font-family: 'Poppins', sans-serif !important;
 }
-.container{
-  
-}
 
 .padding-zero{
   padding-top: 0!important;
   padding-bottom: 0!important;
+}
+.no-padding-top{
+  padding-top: 0!important;
 }
 html, body {
   color: $primary !important;
