+++ src/pages/404.js
@@ -1,6 +1,6 @@
 import React from "react"
 
-import Layout from "../components/layout"
+import Layout from "../components/general/layout"
 import SEO from "../components/seo"
 
 const NotFoundPage = () => (
