+++ src/pages/index.js
@@ -1,14 +1,14 @@
 import React from "react"
-import Layout from "../components/layout"
-import SEO from "../components/seo"
-import "bootstrap/dist/css/bootstrap.min.css"
-import Banner from "../components/banner"
-import Sectionone from "../components/sectionone"
-import Hero from "../components/hero"
 import { ParallaxProvider } from 'react-scroll-parallax';
-import ParallaxBox from "../components/parallaxbox"
-
+import ParallaxBox from "../components/homepage/parallaxbox"
+import Layout from "../components/general/layout"
+import SEO from "../components/seo"
+import Banner from "../components/homepage/banner"
+import Sections from "../components/homepage/sections"
+import Hero from "../components/homepage/hero"
+import "bootstrap/dist/css/bootstrap.min.css"
 import "../components/pointer.scss"
+
 const pointer = document.createElement("div")
 pointer.id = "pointer-dot"
 const ring = document.createElement("div")
@@ -97,19 +97,14 @@ class IndexPage extends React.Component {
   render(){
     return(
       <ParallaxProvider>
-      <Layout>
-      
-      
-        
-        <SEO title="excyted | Creative Studio und Filmproduktion"
-            description="Als eine hybride Werbeproduktion entwickeln wir kreative Lösungen für deine Marke, dein Unternehmen oder dein Projekt. Mit dir wollen wir excyten und deine individuellen Projekte Wirklichkeit werden lassen."/>
-        <Hero/>
-        <Banner/>
-        <ParallaxBox/>
-      
-
-        <Sectionone/>
-      </Layout>
+        <Layout>
+          <SEO title="excyted | Creative Studio und Filmproduktion"
+              description="Als eine hybride Werbeproduktion entwickeln wir kreative Lösungen für deine Marke, dein Unternehmen oder dein Projekt. Mit dir wollen wir excyten und deine individuellen Projekte Wirklichkeit werden lassen."/>
+          <Hero/>
+          <Banner/>
+          <ParallaxBox/>
+          <Sections/>
+        </Layout>
       </ParallaxProvider>
     )
   }
