+++ src/pages/jobs.js
@@ -1,6 +1,6 @@
 import React from "react"
 import SEO from "../components/seo"
-import Layout from "../components/layout"
+import Layout from "../components/general/layout"
 import { Col, Container, Row } from "react-bootstrap"
 import { AiOutlineScissor } from "react-icons/ai"
 import { FaCameraRetro, FaTabletAlt, FaObjectGroup, FaPencilRuler } from "react-icons/fa"
