+++ src/pages/services.js
@@ -1,8 +1,8 @@
 import React from "react"
 import SEO from "../components/seo"
-import Layout from "../components/layout"
-import Clients from "../components/clients"
-import ServicesComponent from "../components/services"
+import Layout from "../components/general/layout"
+import Clients from "../components/sites/clients"
+import ServicesComponent from "../components/sites/services"
 
 const Services = () => (
   <Layout>
