+++ src/pages/team.js
@@ -1,8 +1,8 @@
 import React from "react"
-import Layout from "../components/layout"
+import Layout from "../components/general/layout"
 import SEO from "../components/seo"
-import Values from "../components/values"
-import Members from "../components/members"
+import Values from "../components/sites/values"
+import Members from "../components/sites/members"
 
 const Team = () => (
   <Layout>
